+++ src/pages/ReviewReplyPage/components/ReviewReply/index.tsx
@@ -3,7 +3,7 @@ import { useFormContext } from 'react-hook-form'
 import { Profile } from '@/components'
 import { Data, Receiver } from '@/apis/hooks/useGetReviewFirst'
 import { CheckInTheCircleIcon } from '@/assets/icons'
-import { ReviewReplyType } from '../../types'
+import { ReviewReplyStartType } from '../../types'
 import Questions from '../Questions'
 
 interface ReviewReplyProps {
@@ -12,7 +12,7 @@ interface ReviewReplyProps {
 }
 
 const ReviewReply = ({ reviewData, handleSubmit }: ReviewReplyProps) => {
-  const { getValues, setValue } = useFormContext<ReviewReplyType>()
+  const { getValues, setValue } = useFormContext<ReviewReplyStartType>()
   const receivers = getValues('receiverList')
   const questions = reviewData.questions
